Moves build_dataflow to execution module
